+++ indico/modules/events/abstracts/models/comments.py
@@ -64,7 +64,7 @@ class AbstractComment(ProposalCommentMixin, ReviewCommentMixin, db.Model):
     def can_edit(self, user):
         if user is None:
             return False
-        return self.user == user or self.abstract.event.can_manage(user)
+        return self.user == user or self.abstract.event.can_manage(user, permission='abstracts')
 
     def can_view(self, user):
         if user is None: